Move AbstractFeatureReader.hasBlockCompressedExtension to IOUtil
