+++ src/main/java/htsjdk/tribble/AbstractFeatureReader.java
@@ -18,6 +18,7 @@
 
 package htsjdk.tribble;
 
+import htsjdk.samtools.util.IOUtil;
 import htsjdk.tribble.index.Index;
 import htsjdk.tribble.util.ParsingUtils;
 import htsjdk.tribble.util.TabixUtils;
@@ -57,6 +58,8 @@ public abstract class AbstractFeatureReader<T extends Feature, SOURCE> implement
 
     private static ComponentMethods methods = new ComponentMethods();
 
+    /** @deprecated use {@link IOUtil#BLOCK_COMPRESSED_EXTENSIONS} instead. */
+    @Deprecated
     public static final Set<String> BLOCK_COMPRESSED_EXTENSIONS = Collections.unmodifiableSet(new HashSet<String>(Arrays.asList(".gz", ".gzip", ".bgz", ".bgzf")));
 
     /**
@@ -170,52 +173,28 @@ public abstract class AbstractFeatureReader<T extends Feature, SOURCE> implement
     }
 
     /**
-     * Whether a filename ends in one of the BLOCK_COMPRESSED_EXTENSIONS
-     * @param fileName
-     * @return
+     * @deprecated use {@link IOUtil#hasBlockCompressedExtension(String)}.
      */
+    @Deprecated
     public static boolean hasBlockCompressedExtension (final String fileName) {
-        String cleanedPath = stripQueryStringIfPathIsAnHttpUrl(fileName);
-        for (final String extension : BLOCK_COMPRESSED_EXTENSIONS) {
-            if (cleanedPath.toLowerCase().endsWith(extension))
-                return true;
-        }
-        return false;
+        return IOUtil.hasBlockCompressedExtension(fileName);
     }
 
     /**
-     * Remove http query before checking extension
-     * Path might be a local file, in which case a '?' is a legal part of the filename.
-     * @param path a string representing some sort of path, potentially an http url
-     * @return path with no trailing queryString (ex: http://something.com/path.vcf?stuff=something => http://something.com/path.vcf)
+     * @deprecated use {@link IOUtil#hasBlockCompressedExtension(File)}.
      */
-    private static String stripQueryStringIfPathIsAnHttpUrl(String path) {
-        if(path.startsWith("http://") || path.startsWith("https://")) {
-            int qIdx = path.indexOf('?');
-            if (qIdx > 0) {
-                return path.substring(0, qIdx);
-            }
-        }
-        return path;
-    }
-
-    /**
-     * Whether the name of a file ends in one of the BLOCK_COMPRESSED_EXTENSIONS
-     * @param file
-     * @return
-     */
+    @Deprecated
     public static boolean hasBlockCompressedExtension (final File file) {
-        return hasBlockCompressedExtension(file.getName());
+        return IOUtil.hasBlockCompressedExtension(file.getName());
     }
 
     /**
-     * Whether the path of a URI resource ends in one of the BLOCK_COMPRESSED_EXTENSIONS
-     * @param uri a URI representing the resource to check
-     * @return
+     * @deprecated use {@link IOUtil#hasBlockCompressedExtension(URI)}.
      */
+    @Deprecated
     public static boolean hasBlockCompressedExtension (final URI uri) {
         String path = uri.getPath();
-        return hasBlockCompressedExtension(path);
+        return IOUtil.hasBlockCompressedExtension(path);
     }
 
     /**
@@ -240,7 +219,7 @@ public abstract class AbstractFeatureReader<T extends Feature, SOURCE> implement
         if(indexPath == null){
             indexPath = ParsingUtils.appendToPath(resourcePath, TabixUtils.STANDARD_INDEX_EXTENSION);
         }
-        return hasBlockCompressedExtension(resourcePath) && ParsingUtils.resourceExists(indexPath);
+        return IOUtil.hasBlockCompressedExtension(resourcePath) && ParsingUtils.resourceExists(indexPath);
     }
 
     public static class ComponentMethods{
