+++ src/main/java/htsjdk/tribble/bed/BEDCodec.java
@@ -23,6 +23,7 @@
  */
 package htsjdk.tribble.bed;
 
+import htsjdk.samtools.util.IOUtil;
 import htsjdk.tribble.AbstractFeatureReader;
 import htsjdk.tribble.AsciiFeatureCodec;
 import htsjdk.tribble.annotation.Strand;
@@ -229,7 +230,7 @@ public class BEDCodec extends AsciiFeatureCodec<BEDFeature> {
     @Override
     public boolean canDecode(final String path) {
         final String toDecode;
-        if (AbstractFeatureReader.hasBlockCompressedExtension(path)) {
+        if (IOUtil.hasBlockCompressedExtension(path)) {
             toDecode = path.substring(0, path.lastIndexOf("."));
         } else {
             toDecode = path;
