+++ src/main/java/htsjdk/variant/variantcontext/writer/VariantContextWriterBuilder.java
@@ -502,7 +502,7 @@ public class VariantContextWriterBuilder {
         if (outFile == null)
             return false;
 
-        return AbstractFeatureReader.hasBlockCompressedExtension(outFile);
+        return IOUtil.hasBlockCompressedExtension(outFile);
     }
 
     private VariantContextWriter createVCFWriter(final File writerFile, final OutputStream writerStream) {
